// src/app/interfaces/IPokeapi.ts
export interface IPokemonListResponse {
  count:    number;
  next:     string;
  previous: null;
  results:  Result[];
}

export interface Result {
  name: string;
  url:  string;
}